added some minor styling to build foundation for how the page will be layed out
