+++ react-app/src/components/NavBar.css
@@ -1,12 +1,12 @@
 nav {
-    width: 100vw;
+    width: 100%;
     height: 50px;
     display: flex;
     justify-content: space-between;
-    background-color: #FFFDFA;
+    background-color: transparent;
     align-items: center;
 }
-nav ul:not(.profile-dropdown) {
+nav ul {
     text-decoration: none;
     display: flex;
     flex-direction: row;
@@ -19,12 +19,11 @@ nav ul:not(.profile-dropdown) {
     color: #657394;
     font-family: sans-serif;
     font-weight: bold;
-    margin-left: 10px;
-    font-size: 25px;
+    margin-left: 50px;
+    font-size: 30px;
     border: none;
-    background-color: #FFFDFA;
+    background-color: transparent;
     padding: 5px;
-    margin-right: 10px;
     cursor: pointer;
 }
 .NavButtons{
@@ -32,47 +31,22 @@ nav ul:not(.profile-dropdown) {
     font-family: sans-serif;
     font-weight: bold;
     text-decoration: none;
-    color: #948A73;
+    color: #FFFDFA;
     border: none;
-    background-color: #FFFDFA;
+    background-color: transparent;
     padding: 10px;
-    margin-right: 20px;
+    margin-right: 10px;
     cursor: pointer;
 }
 .signupNav-button{
-    border: 1px solid #948A73;
+    border: 1px solid #FFFDFA;
     border-radius: 10px;
     transition-duration: .3s;
     transition-property: background-color, color;
 }
 .signupNav-button:hover{
-    background-color: #948A73;
-    color: #FFFDFA;
-}
-.profile-dropdown{
-    justify-content: flex-end;
-    width: fit-content;
-    text-decoration: none;
-    list-style-type: none;
-    background-color: #948A73;
-    position: absolute;
-    transform: translate(-50px, 4px);
-    margin: 0;
-    padding: 0;
-    z-index: 99;
-}
-.profile-dropdown > li{
-    display: flex;
-    justify-content: flex-end;
-}
-.profile-dropdown:after{
-    content: "";
-    position: absolute;
-    transform: translate(64px, -80px);
-    border-top: none;
-    border-right: 10px solid transparent;
-    border-left: 10px solid transparent;
-    border-bottom: 10px solid #948A73;
+    background-color: #FFFDFA;
+    color: black;
 }
 .logout-profile-button, .add-listing{
     color: #FFFDFA;
@@ -80,7 +54,12 @@ nav ul:not(.profile-dropdown) {
     border: none;
     cursor: pointer;
     padding: 10px;
+    margin-right: 10px;
 }
-.add-listing{
-    border-bottom: 1px solid #FFFDFA;
+.logout-profile-button, .add-listing{
+    font-family: sans-serif;
+    font-size: 14px;
+}
+.loginNav-button:hover, .logout-profile-button:hover, .add-listing:hover{
+    text-decoration: underline;
 }
