+++ react-app/src/components/NavBar.js
@@ -1,19 +1,37 @@
-import React from 'react';
-import { useSelector } from 'react-redux';
+import React, { useState } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 import { NavLink } from 'react-router-dom';
-import ProfileButton from './auth/ProfileButton';
+import SurfboardForm from "./SurfboardListings/SurfboardForm";
+import { Modal } from '../context/Modal';
+import { logout } from '../store/session';
 import LoginModal from './LoginModal';
 import SignUpModal from './SignupModal';
 import './NavBar.css'
 
 const NavBar = () => {
+  const dispatch = useDispatch();
   const sessionUser = useSelector(state => state.session.user);
+  const [showListingModal, setShowListingModal] = useState(false);
+
+  const log_out = async () => {
+    await dispatch(logout());
+  };
+
+  const callSetter = () => {
+      setShowListingModal(false);
+  };
+
   let sessionLinks;
   if (sessionUser) {
     sessionLinks = (
-      <li>
-        <ProfileButton />
-      </li>
+      <>
+        <li>
+            <button className="add-listing" onClick={() => setShowListingModal(true)}>+New Listing</button>
+        </li>
+        <li>
+            <button className="logout-profile-button" onClick={log_out}>Log Out</button>
+        </li>
+      </>
     )
   } else {
     sessionLinks = (
@@ -35,6 +53,11 @@ const NavBar = () => {
       <ul>
         {sessionLinks}
       </ul>
+      {showListingModal && (
+          <Modal onClose={() => setShowListingModal(false)}>
+              <SurfboardForm callSetter={callSetter} />
+          </Modal>
+      )}
     </nav>
   );
 }
