+++ react-app/src/components/SurfboardListings/ListingsPage.css
@@ -1,14 +1,20 @@
+.main-page{
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+}
 .surfboards-list-page{
     width: 100%;
     overflow-y: scroll;
     margin-bottom: 75px;
+    height: calc(100vh - 75px);
 }
 .surfboards-list-page::-webkit-scrollbar {
     display: none;
 }
 .filters{
     display: flex;
-    width: calc(100% - 321px);
+    width: auto;
     align-items: center;
     position: absolute;
     transform: translate(calc(50% - 100px), 50px);
@@ -45,17 +51,13 @@
     margin: 125px 0 0 20px;
     padding-right: 40px;
 }
-.listings-list > li{
-    height: 125px;
-}
 .individual-item{
     display: flex;
-    flex-direction: row;
+    flex-direction: column;
     padding: 20px;
-    border: 1px solid #C5CEE0;
-    border-radius: 20px;
+    border: none;
     text-decoration: none;
-    background-color: #FFFDFA;
+    background-color: transparent;
 }
 .listing-img{
     padding: 10px;
@@ -64,7 +66,7 @@
     width: 120px;
 }
 .listing-img > img{
-    height: 59px;
+    height: 180px;
     width: 100%;
 }
 .listing-img > i{
@@ -81,25 +83,8 @@
     font-size: 30px;
     font-family: sans-serif;
     color: black;
-    margin-left: 20px;
 }
 .listing-size{
     font-size: 18px;
     color: black;
-}
-.listing-description{
-    margin-left: 20px;
-    font-size: 16px;
-    color: black;
-}
-.listing-info{
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    width: 100%;
-}
-.locationANDsize{
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
 }
