+++ react-app/src/index.css
@@ -2,10 +2,11 @@
 body{
     margin: 0;
 }
-.main-page, .listings-page{
+.whole-page{
     display: flex;
-    width: 100%;
-    height: 100%;
+    flex-direction: row;
+    width: 100vw;
+    height: 100vh;
     position: fixed;
     background-image: url('https://hi-surf-dev.s3.us-west-1.amazonaws.com/BackgroundIMG.jpg');
     background-size: 100% 100%;
